refactor(navbar): centralize navigation links and move getInitials util
- Extract NAV_LINKS constants into a dedicated nav-config module
- Replace inline nav links with imports from nav-config in desktop and mobile nav
- Merge USER_MENU_LINKS with APP_NAV_LINKS for authenticated nav items in mobile nav
- Remove duplicate getInitials functions and consolidate into shared utils module
- Update UserMenu and mobile nav to import getInitials from utils
- Simplify nav links memoization by returning predefined constants based on auth state
