+++ libs/ui/src/lib/UserMenu.tsx
@@ -10,6 +10,7 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from '../components/ui/dropdown-menu';
+import { getInitials } from './utils';
 
 interface User {
   name: string;
@@ -21,13 +22,6 @@ interface User {
 interface UserMenuProps {
   user: User | null | undefined;
   onLogout: () => void;
-}
-
-function getInitials(name?: string): string {
-  if (!name) return '';
-  const parts = name.trim().split(/\s+/);
-  if (parts.length === 1) return parts[0].substring(0, 2).toUpperCase();
-  return (parts[0][0] + parts[parts.length - 1][0]).toUpperCase();
 }
 
 export const UserMenu = ({ user, onLogout }: UserMenuProps) => {
